requirements.txt
flask==3.0.2
flask-cors==4.0.0
langchain==0.1.9
langchain-community==0.0.24
langchain-openai==0.0.8
openai==1.12.0
python-dotenv==1.0.1
chromadb==0.4.22
pdfplumber==0.10.3
python-docx==1.1.0
tiktoken==0.6.0
